Modularize: Export core's `Locale` from our Locale
This lets downstream consumers import this file to
guarantee that they add our strings.

This also fixes bugs when downstream consumers end
up with multiple copies of core or hdate, in which
case they would have had no way to fetch a copy of
`Locale` that includes both `core` and `leyining`.